Refacto ReferralDetails.tsx
