+++ src/frontend/js/const/index.ts
@@ -1,0 +1,9 @@
+export enum nestedUrls {
+    answer = 'answer',
+    content = 'content',
+    draftAnswers = 'draft-answers',
+    draftAnswer = 'draft-answer',
+    messages = 'messages',
+    tracking = 'tracking',
+    users = 'users',
+}
